feat(getEntries query string):getEntries query string api endpoint
create search entry method inside entry controller class
add route for search entry

[de3liver #160195369]

+++ client/public/js/delete-entry.js
@@ -1,25 +1,30 @@
 const deleteEntry = () => {
-  const message = document.getElementById('success-message');
-  const close = document.getElementById('close');
-  // document.getElementById('loading').style.display = 'block';
+  const confirmDelete = confirm("Delete this entry?");
+  if (confirmDelete) {
+    const message = document.getElementById('success-message');
+    message.style.display = 'block';
+    const close = document.getElementById('close');
 
-  const option = {
-    method: 'DELETE',
-    headers: {
-      'x-access-token': token,
-      'Content-Type': 'application/json',
-    },
-  };
-  fetch(`https://your-diary.herokuapp.com/api/v1/entries/${entryId}`, option)
-    .then(res => res.json())
-    .then((data) => {
-      message.style.display = 'block';
-    })
-    .catch((err) => { console.log(err); });
+    // document.getElementById('loading').style.display = 'block';
 
-  close.onclick = function () {
-    window.location.replace('./entries.html');
-    message.style.display = 'none';
-  };
+    const option = {
+      method: 'DELETE',
+      headers: {
+        'x-access-token': token,
+        'Content-Type': 'application/json',
+      },
+    };
+    fetch(`https://your-diary.herokuapp.com/api/v1/entries/${entryId}`, option)
+      .then(res => res.json())
+      .then((data) => {
+        console.log(data)
+      })
+      .catch((err) => { console.log(err); });
+    close.onclick = function () {
+      window.location.replace('./entries.html');
+      message.style.display = 'none';
+    };
+  }
 };
-// document.getElementById('delete').addEventListener('click', deleteEntry, false);
+
+document.getElementById('delete').addEventListener('click', deleteEntry, false);
